IOF-750 cart: Create a CMF for file names rather than using strings.
Embed 256 character arrays in input RPC types rather than using strings.
Add a new CMF_IOF_NAME type, and use this for RPC inputs which take a
filename.

This increases the amount of data transfered in RPCs so that it's always
the same size, however on the flip side it means all data is inline,
there is no additional memory allocation and all data will be contiguous
in memory.

Change-Id: I70688799f3f962fb1936b3ada74c27369218e603
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30088
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>