IOF-763 threading: Update IOF threading code.
Create a iof_ctx datastructure to contain the common elements
needed to start and stop the per-ctx progress thread.  Embed
a copy of this in iof_state and update code accordingly.

Use a signal handler in the FUSE code to prevent deadlock if
the FUSE thread is sleeping when we request a shutdown.

Change-Id: I04c3314e94f655058b7a219b87ea5930c7417bb3
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30178
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>