IOF-705 fuse: Do not set filesystem offline before attempting shutdown.
Do not set the filesystem offline in IOF before unmounting it but rather
allow the kernel and fuse itself to handle the unmount procedure.

This helps because the kernel keeps many handles open and if we disable
access before unmounting we then reject requests to close these handles
and shutdown becomes a problem.  Instead simply let the kernel deal with
it and treat all requests that arrive at IOF equally.

Change-Id: I44d0a183d8557c9d8934acf6dc7d49a23c2f77e8
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30244
Tested-by: Maloo <hpdd-maloo@intel.com>