IOF-765 test: Fix a bug in test teardown.
Correctly wait for a time after requesting shutdown before
asserting that processes have exited.  The tests were all
passing before this so obviously 30 seconds was to long
so decrease the time to 10 seconds however actually wait
this long.

This appears to have been the cause of sporadic test failures
in HA for a while.

Change-Id: I855e9259b0653a25a9a4c7faaafba263a9bf35a0
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30258
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>