IOF-343 testing: Update common_launch to take array.
Update common_launch_(cmd|process) to take an array rather
than a string as this removes an intermediate step.
Remove the first option as it was mostly mis-used and instead
use the first array element.
Remove the rlimit code as the ionss will call this
Remove the old_test directory entirely.

Change-Id: I20a5212e72a979dbdbe5150c58c4db944685ab81
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30439
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>