IOF-773 rpctrace: Streamline the rpctrace logging code.
Rather than open and close the log file for each line of logging
open the file once at startup and pass the file object into
the rpc tracing code and have that write to a file descriptor.

This both makes the code slightly clearer, but also means that
the file isn't opened and closed for every write meaning much
less work to do on the backend.

Change-Id: I59468b6b85ff29ce0d93d9f9888a9455716b5fdd
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30459
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Sydney Vanda <sydney.m.vanda@intel.com>