IOF-739 getattr: Use information in fi if available.
If there is a file handle associated with a request then use it
in preference to doing a inode lookup.
This means avoids a hash table lookup for open files but also means
that we do the server-side operation on the file handle rather than
the inode handle, and can also avoid doing to operation of the GAH
has been reported as invalid.
Update setattr to be similar.

Change-Id: Iac388e2733f52ea98e9d05b4b49b4f36adef876a
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30498
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Omkar Kulkarni <omkar.kulkarni@intel.com>