IOF-781 pools: Use a handle per pool rather than per type.
In almost all cases the "handle" arg was the same for every type
in a pool so enforce this and make the handle an attribute of the
pool rather than the type.  This simplifies the object hierarchy
somewhat, and makes it easier to trace the logs whilst being
a cleaner interface.

Move to having a pool per projection in the ionss.  This means
slighly more overhead however each type was registered per
projection anyway so really it just makes the existing
distinctions clearer.

Change-Id: I185b35bc7b65131601c87a982822652ea64db6c1
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30621
Tested-by: Maloo <hpdd-maloo@intel.com>