IOF-343 ioc: Update return codes for thread start|stop.
Thread start now returns a bool only indicating success.
Thread stop returns the code from crt_context_destroy().

Remove the use of rcb (return-code-boolean) variables in many cases
and just test the return of the function directly without storing
it to a variable if that variable is never reused.  This removes
several variables, at least one of which was a int rather than a bool
and several were scoping violations.

Change-Id: Iab0ab463373f89ae9adcb66f4758ab3b9d728f14
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30894
Reviewed-by: Omkar Kulkarni <omkar.kulkarni@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>