IOF-767 ionss: Do not resubmit RPC on corpc failure.
The code here was simply wrong, it is not possible to re-submit a RPC
if it fails, there is a requirment to create a new RPC in this case.

This changeset simply removes the errant code, and simply logs on
failure, it does not attempt to create a new RPC in this case.

Change-Id: I9744020680b35cdb04158161a6aa326559badeab
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30920
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Omkar Kulkarni <omkar.kulkarni@intel.com>