IOF-798 cnss: Update reference counting for RPCs.
In the write code take a reference before submitting a RPC.  Previously
we thought it was safe to take this reference after submitting the RPC
however we have seen the RPC complete, the descriptor be freed and re-used
before the req_send() function returns, which leads to an assertion
failure in CaRT itself.  To avoid this always take a reference before
calling req_send() rather than after.

Having taken a reference before req_send() be sure to release it again
should the req_send() call fail.

Change-Id: If957b622eea22d09f4852309259dfa0d89142c23
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30921
Tested-by: Maloo <hpdd-maloo@intel.com>