IOF-802 readdir: Fix check for buffer-full from dentry callback.
Correct an error when populating directory contents, and correctly
stop supplying data to FUSE when the buffer is full.  This means
that we do not consume extra resources on entries that won't be used
and future callbacks can continue from the current location in the
directory stream.

Change all IOF_LOG() calls to use IOF_TRACE()

Change-Id: Ia6b533ae04ca3d2bcb8e3b9123e2af24f592a92b
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/30939
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Omkar Kulkarni <omkar.kulkarni@intel.com>