htable: Make the inode hash table size configurable.
Add a inode_htable_size option to projections to control the size
of the hash table.  This is used on both the CNSS and IONSS.

The default value is 5, changed from 4 on the CNSS and 5 on the ionss.

Change the compile order of the ionss around to give better compiler
errors when adding new options.

Change-Id: I5fc3ef5713575b2d31da46246bdd5e6bf956a69c
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/31011
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Omkar Kulkarni <omkar.kulkarni@intel.com>