IOF-812 stats: correctly record the read_bytes stat.
There was a bug where this value was being updated from rc
rather than read_bytes and as rc was always zero no stats were
recorded.  As the rc variable use to be a errno of negative or
byte count if possitive this code was correct however did not
get updated correctly when the bytes_read variable was added.

Change-Id: I6beb55c8e1b488148ebc3d97ac3d8ecd4f5544b8
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/31285
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>