IOF-96 doxygen: Initial doxygen work, including GAH docs.
Write a doxygen configuration file.  To do this I created a template
file and then edited to suit, commenting my edits.

Update the ios_gah code to be in the format doxygen expects.
A lot of the code in IOF is using "doxygen like" comments but is
in many cases doing so incorrectly, often mixing before and after
comment identifiers so the comment is applied to the wrong entity.
Fix this in ios_gah.h and make use of the \retval tag correctly as
well.
Annotated comments within c functions were being added to the
function documentation which in many cases was not intended so
remove the annotation here.

Doxygen uses both brief and detailed annotations but our code mostly
just defined detailed so enable the AUTOBRIEF setting to extract
the first sentance from detailed comments to be the brief comment.

To run this use the following command, output will be saved in the
'autodocs/html' directory

$ doxygen utils/doxygen.conf

Change-Id: I1dfdc5ecc560c4b574355a6005b5239dc8e0b928
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/31570
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>