IOF-756 cnss: Use atomics for checking GAH validity.
Update the "gah_valid" field in file and directory handles so
that it's atomic.  Provide macros for accessing this field.

For file handles this meant moving the bit from the common
field to the handle itself which means it's not accessible
to the IL any more however failover in IL needs to be handled
differently anyway, and we could add the current macros to
the IL at a later date if it makes sense.

Change-Id: I32c89275e66bcca74a82695e6b278d791db0f22b
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/31584
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>