IOF-840 gah: Check root in get_info() call.
Add a check that the root is correct in the get_info() call so
if a request is sent to the wrong node it's correctly detected.
Modify the GAH API to pass a root to the init() call and remove
both root and base from the allocate() call, create a new
allocate_base() function where a custom base is required.
This allows the root to be stored in the struct, so it can then
be checked on calls to get_info() and the IONSS can correctly
reject GAH that it didn't locally allocate.

Change-Id: I0f984c697dd26439f52de0e56651248ce826d2ec
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/31685
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>