IOF-841 testing: Start adding a test to test failure conditions.
Add a first test to check failover.
Setup IOF to have two projections, one with failover enabled one not
Kill a single IONSS process acting as the PSR.
Check that the filesystem not configured for failover returns EHOSTDOWN

Set a defatult timeout to reduce runtime.

Note that this test is disabled by default as it does not yet pass,
and is likely blocked on CaRT doing a OFI update before it can pass.

Change-Id: Ie39e5e2b0536265250a4216d08159d926c19d135
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/31693
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>