IOF-821 failover: Send reregister RPC on failover.
When evicted from a process set send an a RPC to the new PSR so
that it can do reference counting on clients.

Fixup client so that correct return codes are passed on eviction.

Improve test, to give better logging, run test twice (one for
RPC-before-eviction, once for RPC-after-eviction) and check the
return code is as expected in all cases.

Change-Id: Ia58eeecf2b262feee339c028898b1e70ac2b36a0
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/32023
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>