IOF-847 testing: Enable failover_stat test and add valgrind options.
Remove the annotation to skip the test, and remove the check
that pids were contigous.  Skip the test if valgrind if IONSS is
running under valgrind.

This should allow the test to run properly, in corner cases of
pid-wrap etc it may cause false-negatives however.

Add a new option for cnss-valgrind to allow the use of valgrind
on the CNSS only.

Change-Id: If619086324485d61f445036fea58fa67420f0e6a
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/32053
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>