IOF-769 test: Add a --fixed-path option to test_local.
This tells the test code to use a pre-determined path for testing
rather than a temporary directory.  This is useful for running
test_local inside bigger scripts that then know where the projection
is going to be mounted.
If using fixed-paths and there is a stale mount then clean up the
previous mount before starting.  This should avoid development
systems being left with multiple stale mounts.

Change-Id: I51db0016b2c48a2ccd2f85c31ed857dac31ddc53
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/32132
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>