IOF-830 cnss: Propagate error from dereg_fuse() callback.
Allow errors to be returned from deregister_fuse callback to main,
and exit non-zero if errors are returned.  Fix a pthread_mutex_lock()
that was failing, and logging an error but not failing any tests.

This change propagates the error (and fixes the current error) but
would not cause the error to fail the test because of the
--continous option to orterun that masks process failures.

Change-Id: I252e200ea7f81d04b9d7f65f0034db9e99ba820a
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/32448
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>