IOF-323 cart: Add configure check to support latest cart version.
Add a configure check for d_log_sync_mask so that IOF can work with
both the current and latest version.
Update scons_local to latest.
This does not update IOF to a different CaRT but fixes compile issues
so that IOF can be tested with later versions of CaRT.

Change-Id: Ic99c66abc35b7ac606d17449f4b9c983dad212d3
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/32580
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>