IOF-343 testing: Change failover testing from usr to exp.
Change the failover testing to test the exp rather than the usr
projection.  As the usr projection is read-only it's not possible
to use this for r/w tests.
Change the failover setting from one projection to the other.
Change the failover_fstat test to pre-create a file in the export
dir and then open it in the import dir rather than finding a
pre-existing file in /usr.
Split the failover_stat case into two by creating a new test
failover_off_stat for projections where failover is disabled.  In
this case create a helper function which takes a fsid and a list
of expected error numbers to get back from stat operations.

Note that so far this does not increase testing, or add any r/w
testing but allows such tests to be added in future.

Change-Id: I2d4fa5b48ba23a9c4b361b61308327f4d0586747
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/32716
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>