IOF-825 failover: Construct tree of open files on failover event.
When there is a failover event on a projection create a tree
structure of inodes/files based on last-observed information about
the filesystem layout.  For each open file add it to a per-inode
list, for each inode add it to a per-inode list of inodes, so that
the filesystem hierarchy can be traversed.
Note that inodes which do not have any open files or directories
are invalidated in the kernel so do not particiate in this tree, and
the information is based on last-observed state so may be out of date.

Change-Id: I2571b1d862ee8a243216a487e96ef6dcbe374fd4
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32720
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>