IOF-343 test: Increase test shutdown timeout.
We are seeing occasional timeouts shutting down IOF after testing,
it appears that successful runs take approx six seconds and the
timeout was set to 10 so increase the timeout slightly to 20.

Improve the logging and documentation as has_terminated() returns
a bool but it was being printed as a integer.

Change-Id: I6cc390d09141f350445a8379321f4b465093d785
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32790
Tested-by: Maloo <hpdd-maloo@intel.com>