IOF-790 tracing: Remove dupliate rpc_reporting() function.
The functions rpc_reporting and rpc_reporting_multiprocess were
almost identical so replace them with a single function
rpc_reporting_pid() and use rpc_reporting to decide what options
to invoke it with.
As part of this add a set_input_file() method which can be used
on the object to pre-load the pids from the file.

Change-Id: Ifecb99224d730cea47a32f961052ee1d76883218
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32870
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>