IOF-825 cnss: Add gah_ok status to inode handles on client.
The gah_ok flag and macros for accessing it are already used for
file and directory handles, add the functionality to inodes as well.
Update find_gah() to check the gah_ok value before returning the GAH,
and change it to return either ENOENT if not found or EHOSTDOWN if
found but invalid.
Set the gah_ok flag to be valid before adding to the hash table in
both cases.
Update all callers of find_gah() to use the rc returned, rather than
ENOENT in all cases.
Set gah_ok to false in the failover code.

Change-Id: I215058c212d5fe73a503637955a8d883757df758
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32896
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>