IOF-343 cart: Use D_MUTEX_INIT() macro over pthread_mutex_init()
Update all but one case of mutex_init to use the cart macro which
will log on failure.  The one remaining case is in code which
should be removed or reworked soon, and the return code is currently
not checked anyway.

Change return code of iof_pool_init from libc errno to cart error
number.

Change-Id: I6aa9ef58ab33bf998dcd1b154afe5e51d1d4fbf7
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32906
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>