IOF-343 cnss: Clean up some cnss error paths on startup.
Rename list_iter variable to be entry.
use d_list_for_each_entry_safe() rather than d_list_empty in
a while loop.
Destroy the plugin if the plugin dir cannot be created.
Update iof_thread_stop() to make it work even if there is no
thread, as this allows the pluging to be destroyed on partial
init.

Change-Id: Ia45fa14c10f10e09735ab69ef84ecd19cc9f5808
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32909
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>