IOF-825 failover: Add a on_presend callback to requests.
Add a new on_presend callback to be called prior to every send
allowing for the GAH to be re-loaded on resend.

Rather than calling find_gah() to load the gah without taking
a reference call new find_inode() function that returns a pointer
with a reference held and then use this in the presend function.

As the request now has information about the handle use the GAH
to identify the target rank if available.

Improvements to logging so the right handle is pointed in some places.

Split H_GAH_IS_VALID macro into two, one for files and the other for
dirs/inodes and add logging when invalid GAH is encountered.

Change-Id: Icd84652cd0e1b94603e9370ca2e41a1e3b9c6332
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32914
Tested-by: Maloo <hpdd-maloo@intel.com>