IOF-876 cnss: free fuse_ops after initial use and add debugging.
fuse takes a internal copy of the fuse_ops struct so free this after
successful init call, previously we were keeping this around until
just before fuse last uses it's copy.  Initially this appeared a
use-after-free issue however as fuse does take a copy it's a simple
cleanup.

Add a fuse destroy callback which does logging only (and is called
after fs_handle is freed), and add logging around the call to
fuse_session_destroy() as this function is deadlocking in some cases.

Change-Id: Ia6952a7a648b78af7e293f2412b3b31e0bcde83d
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32941
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>