IOF-788 internals: Convert more cases to use self.lf
Add a level=<value> option to .reset() to only return line above
a specific log level.
Add a to_str() method to lines to return a true string.  to_str()
takes a mark option to add highlight markers to the line.
Fix a couple of spurious end-of-line spaces in the logging.

Change-Id: Ifbf2ff66fbda7f0fea7b833247936100bdf70dc9
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32945
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>