IOF-790 testing: Add a option to test_local to launch instance for each test.
Add a --instance-per-test option which causes test_local to create a new
IOF instance for each test.  This is how the test-runner in CI runs the
tests however test_local run directly normally uses the 'go' method to
run all tests in a single instance.

The advantage of using a instance per test is that you get a set of unique
logs for each test, at the cost of run-time.

Add a skip-failover option to speed up local testing time for trivial changes.

Change-Id: Ief1f4fe6f2caddaefd8fd5d6c09a923cf972186e
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32948
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>