IOF-827 cnss: Add a testcase for unlink after failover.
Try unlink after server failover to verify behaviour.  Unfortunately
the kernel translates a unlink request into lookup+unlink which is
what the CNSS sees, and as lookup is idempotent this can just be
re-sent after failover and succeed, so the request we want to test
(the unlink request itself) is not actually covered by this test.

Add logic to lookup to resend on failure, as currently this code
only existed for fgetattr.

Change-Id: Ib7854953cabed34c4985f6f83f6b98b0b7600e89
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32949
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>