IOF-877 cnss: Update setattr to use request API.
Refresh the getattr code to be more generic.
Update the setattr code to use request api and match the getattr code.

Rework the common_req init code now there are three users of it to
work in the same way as the entry_req init code where there is one
init function which sets the opcode, and a single reset function
for all pools.

Change IOF_FUSE_REPLY_ATTR to IOC_REPLY_ATTR and have it take a request,
in time all the IOF_FUSE macros will need to change in order to be
consistent and avoid race conditions in the logs.

Change-Id: Ifc2e038f6089e9258d98ce1a0e78f57fb067ffc6
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32954
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>