IOF-788 interals: Use object for descriptor search and pre-load logs
Read the log contents once into memory, storing it into an array, and
have the iterator read from the array rather than the file.

Update the _find_reused_descriptor_index() function to take a log object
to scan through, rather than opening the file each time.  This means
that the search is performend in-memory rather than from disk each time.

Add a "index" to the log lines, and the iterator so that iteration can
start from a specific location, rather than traversing the file looking
for the starting location.

This gives a significant speed improvement, and also addresses an issue
where duplicate lines in the log file could lead to incorrect descriptor
reporting.

Change-Id: I6b39243978c655801f7bf6c44fbcf4b02e7f34fc
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/32991
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>