IOF-788 internals: Improve internals tracing code and output.
Improve the logic for tracing descriptors whilst iterating over the file,
use a OrderedDict for descriptors so they can be displayed in time order.
Keep a reuse_table so that reuse counts are tracked rather than being
recalculated each time they're needed.
Combine the loops for linking and using the descriptors, so the reuse
count can be read at use, rather than having to re-read the entire
file for each time a pointer is reused.
Add a tabulate_output() method for showing a table via tabulate
directly.
Inline the _rpc_tabulate() function
Combine the RPC and descriptor tables in the output, and use tabulate
to make them align properly.  Becuase of the OrderedDict they are now
shown in time order as well.

Change-Id: I5e74c59f05f0467e3a7666eceb5342948a9fb303
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33002
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>