IOF-788 internals: Try and reduce memory usage in internals code.
Having recently tried (and failed) to use internals tracing on a
250Mb log file try to reduce some of the memory usage in logparse.

When putting the log in the internals file do not buffer the whole
thing in memory, but rather write each line at a time.
When reading from file do not save self.fields, but calculate on demand
use rstrip() rather than strip() for removing the \n at the end of a line.
Fix a bug in to_str() relating to line lengths, as the pid is printed in
the output the column offsets change based on this.

Change-Id: I8e5514f49f3751d387958da79ccf69cd3c1a6f32
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33009
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>