CART-880 cnss: Delay dentry invalidate until after release lock.
To avoid deadlock in the kernel do not call inval_entry until
after dropping the gah_lock.  Instead of calling inval_entry()
in the hash table traverse, use the traverse to populate a list
of inodes to be invalidated, and walk this list later on in the
failover process once the gah_lock has been dropped and fuse
requests from the kernel are being processed again.  This avoids
a deadlock where the kernel is trying to perform an operation
on the parent and the client is trying to perform an operation
on the child.

Rename some entries in the inode_entry struct for clarity and improve
the comments.

Change-Id: I7ceaaaaf2714ad6d9a625b747bbd4620d23a5191
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33028
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>