IOF-877 cnss: Use the new IOC_ENTRY macros for lookup.
Update the lookup/mkdir/symlink code to use the new IOC_REPLY
macros.  This means that the IOC request is used as the logging
handle rather than the fuse request, so removes the race
condition when logging the result of the reply() function, after
the fuse descriptor has been released.

In the internals tracing code only look for descriptors created
in fuse code, not used there to ensure handle is valid.

Change-Id: I941128d7b027debcae67f37fec91469c479fbcee
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33062
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>