IOF-877 cnss: Change warning message to info only on shutdown.
Change the message level from WARNING to INFO when the fuse session
is still active on the first shutdown, the process is the code
requests the thread to stop, but the thread has to first notice
the request, and then do a lot of cleanup before it exits.  If the
thread is slow then a warning message was being printed and
the internals tracing code was picking this up, and using it as
a descriptor to trace, when in fact there was no actual error
to trace.

In addition, the handle passed to the logging calls in this
function was not valid, so change all logging sites to use
the plugin, rather than the info pointer.

Change-Id: Ib6cf5ea4ec2f8e5f73a25ebd8465948c4714c9c9
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33063
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>