IOF-845 cnss: Remove the on_presend callback and replace with offset.
All the on_presend functions are essentially the same, so replace all
on_presend functions with a gah_offset value, and in-line the GAH
copy code in the ioc_fs_send() function.
Remove the in arg to IOC_REQ_INIT_REQ() as it sometimes wasn't used
so define and set in only where needed.
Add a RHS_DIR type of handle, and support extracting the GAH from
directory handles.
Remove the fs_online and GAH_VALID checks from the callbacks, and
handle this entirely in ioc_fs_send()

Change-Id: I0d601ad11bd168f72b6759dbd83af83fe45ba7b2
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33107
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>