IOF-343 test: Tell orterun not to try and use openib.
Add the options --mca btl tcp,self to orterun so that it won't try
and access the IB network.  This should be used only by the orterun
binary itself, and possibly PMIX but not by CaRT.

Rework the IofRunner module to build the command as an array rather
than as a string and converting it.

update scons_local to pull in the fix there for orte-dvm.

Change-Id: I150acab41160eb039daa2d259d10ddcf62b21609
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33200
Reviewed-by: John Malmberg <john.e.malmberg@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>