IOF-882 cnss: Remove on_send callback entirely.
The use of on_send was unstable as the request could have been freed
by the time the callback was used.  Remove the callback entirely.

This will have a performance impact, on_send was used to pre-populate
descriptors for future use, so this change will mean some calls now
do allocation on the critical path, whereas before most would not.

This could potentially be avoided by adding the pool pointer into
the request itself, and instead of calling on_send call restock()
directly with the pool pointer, however this would mean putting a
pointer in API, which cannot be done statically, or putting a pointer
in request itself.

Change-Id: I00edbad6d18eb48d102676af00290bc574a9095d
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33207
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>