IOF-790 logging: Improvements to the log checking code.
Only report each relationship error once.

When checking for trace relationship errors only log each error
message once, and try to remove unique values in the output so
that messages are less likely to be unique.

Log where lost descriptors are created, by saving a copy of the log
line and using then at shutdown.

Track memory allocations/frees and log on errors or leaked memory here.

This helps with source-line reporting of messages in CI, as it
reduces the messages reported.

Change-Id: I62eef3478f7bdd94b59ce080f8f66007d507e0d6
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33217
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>