IOF-790 internals: Improve the iterator code in internals tracing.
Use iterators properly, and remove the reset() method, replacing
it with a new_iter() method which returns an iterator so that starting
an iterator is now a one-step rather than a two-step process.

Clean up some '{0}'.format() code where the index was not required.

Use line.parent and line.desc where possible, and improve documentation
here.

Move some methods around in the source code so that methods which are
helper methods only are immediately above where they are invoked from.

Use a truth-table for state transitions rather than a series of if
statements.

Change-Id: If2d1b70ac6c544f1adb139f3eb1d2ebc619e64bc
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33222
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>