IOF-731 cnss: Update open to use requests API.
Embed a ioc_request in the file handle, rather than just the RPC and fuse
request.  This will allow open to succeed after failover.

As part of this remove the fs_handle pointer from file handle as it's now
duplicated with open_req.fsh so update a lot of code to use the new
location for this handle.

Change-Id: I57a5e10c278bdbedbd612c00f8e454e9ded8675f
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33245
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>