IOF-669 ionss: Check for trace consistency in ionss log files.
Check the ionss log file for consistency, in the same way the cnss
log file is checked.  This means adding code to iterate over pids
in the log file as each ionss process saves to the same file.

Fix errors with trace handles, creating handles for root, projection
inode handles and directory handles.

CaRT is logging some allocations with different masks to the
corresponding free calls so turn on PMIX and GROUP debugging
or else these get reported as errors.

Change-Id: Ib8008163f324227788639bcf3271afefdcac14d1
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33267
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>