IOF-885 config: Add a cnss_timeout option for projections.
Change from a hard-coded value of 5 seconds for CNSS timeout to
a default value of 60 seconds, but make it configurable.  Have the
test code set the value to 5 seconds.

Improve the option parsing so that the ionss exits rather than ignores
unknown options.

Add a missing CMF_UINT32 to the iof_psr_query so it's also possible to
set the poll interval as well.

Change-Id: I919c63dc2f0138f881014690354e59238b850b71
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33386
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Jenkins