IOF-888 cnss: add some hash table debugging and error checking.
Check the return code of d_hash_rec_ndecref(), and log if it fails,
promoting the log message to a test failure rather than silently
ignoring it.  This should help identify any occasional issues.

Improve the debugging for inodes, by using the inode pointer rather
than fs_handle in log messages in a lot of places.

Add a test that accesses a file by two names, and causes the kernel
to clobber an old inode as this is the cause of at least one
reference counting bug.

Assert that the reference count is zero when closing an inode, and
set the initial reference count to 1 as part of close, so that if
the descriptor is re-used it's correctly zeroed at the start.

Change-Id: I0fa1a1bd8a0d3e7e7c0bbef41c6962badef135a2
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33424
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Jenkins