IOF-887 ctrlfs: Wait for ctrl_fs thread to start before proceeding.
During ctrl_fs_start() wait for the helper thread to be invoked,
and the .init callback to be called before progressing.

It appears that on occasion this has been very slow to be scheduled,
leading to race conditions and apparent memory leaks when the process
makes an early exit, for example in the fault injection tests where
the CNSS process is very short lived.

Change-Id: I949c9a8ff18ef92750b463154d85cfb8a1e50308
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33438
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>